♻️ Begin refactoring main types out of lib.rs
✨ Add basic `impl` for `Aberration<M, F>`
📝 Cleanup module documentation

The amount of documentation in the main library file makes it hard to
navigate the source code itself.

At the moment we're moving the Aberration and Concern types into a
submodule named 'stable', however we'll most likely separate each one
into their own file and then place either nightly or unstable features
in each respective source file.

The layout of this project is atypical, but this project itself is
atypical so I can't say I'm surprised to be honest. 😅

Signed-off-by: Isabella Muerte <63051+slurps-mad-rips@users.noreply.github.com>