🚧 Experimenting with Outcome and color-eyre's Section
This set of changes is currently experimental and a work in progress. If
it *does* get merged it will do so with quite a few more changes and
obviously more code samples.

This Section support/reimplementation is locked behind a feature gate,
pretty-report. This name will most likely change prior to merging, as we
also need to reorganize the entire crate at this point. The number of
cfg_attr and cfg(feature = ...) is starting to get out of hand.

Signed-off-by: Isabella Muerte <63051+slurps-mad-rips@users.noreply.github.com>