📝 Add initial documentation on state escalation API (#13)
* 📝 Add initial documentation on state escalation API

✨ Add initial set of state escalation calls
🐛 Fix missing docs when creating a crate

It should be noted that as of right now, we need to worry about the
situation where a user might not remember to enable all features for a
given feature. For this reason we'll need to, at some point, refactor
the nightly module and create a few more subfeatures. This will break up
the `nightly` feature a bit more, but on the bright side the `nightly`
feature becomes an umbrella feature.

Lastly, despite these initial docs regarding state escalation we don't
have examples in place (nor tests!) to support our interfaces. This will
be rectified with later PRs, most likely after the nightly feature
refactor.

Signed-off-by: Isabella Muerte <63051+slurps-mad-rips@users.noreply.github.com>

* 🚨 Resolve lint checks

Signed-off-by: Isabella Muerte <63051+slurps-mad-rips@users.noreply.github.com>