👷 Hacks to just get this working
More work is needed to properly support caching these tools and doing it
in shell scripts is *just* not feasible.

For now, downloading latest and untarring everytime should do.

NOTE: We did run into a situation where a macOS runner already had the
file downloaded! We skip that now if its present

Signed-off-by: Isabella Muerte <63051+slurps-mad-rips@users.noreply.github.com>